Move ServiceBus creation to regular object instantiation.
